Treating empty XFCC header as a **validation error**.
It's not an extraction error (there is nothing to extract). In cases
where the header is required it makes more sense as a 401 (validation
error) than as a 400 (extraction error).

Also updating the `godoc` in the process to reference the actual
arguments in the signature. This is especially useful because the
acronym `cip` does not match up with the type `IdentityProvider`.